feat(scanner): add robust OCR extraction and verification fallback logic
