+++ apps/web/app/[locale]/scan/page.tsx
@@ -2,99 +2,26 @@
 
 import { useState, useCallback, useRef, useEffect } from "react";
 import { SkeletonLoader } from "@/components/scanner/SkeletonLoader";
-import imageCompression from "browser-image-compression";
-import {
-    Camera,
-    ShieldCheck,
-    Info,
-    AlertCircle,
-    Layers,
-    Copy,
-    Check,
-    Home,
-    Share2,
-    XCircle,
-    AlertTriangle,
-    Search,
-    X,
-    ScanLine,
-    History,
-} from "lucide-react";
+import { useMedicineVerification } from "@/hooks/useMedicineVerification";
+import { VerifiedSafeResult } from "@/components/scanner/results/VerifiedSafeResult";
+import { CounterfeitAlertResult } from "@/components/scanner/results/CounterfeitAlertResult";
+import { UnverifiedResult } from "@/components/scanner/results/UnverifiedResult";
+import { ErrorResult } from "@/components/scanner/results/ErrorResult";
+import { formatExpiryForBadge } from "@/lib/medicineDateUtils";
+import { useMedicineImageUpload } from "@/hooks/useMedicineImageUpload";
+import { Camera, Layers, Search, X, ScanLine, History } from "lucide-react";
 import { Link } from "@/i18n/routing";
 import { PageHeader } from "../components/PageHeader";
 import { toast } from "sonner";
-import { ExpiryBadge } from "@/components/scanner/ExpiryBadge";
-import {
-    verifyMedicine,
-    fuzzyMatchBrand,
-    verifyMedicineByBrand,
-    checkLasaConflicts,
-    type VerifyResult,
-    type LasaMatch,
-    type VerifiedMedicine,
-} from "@/lib/api";
+import { verifyMedicineByBrand, type VerifiedMedicine } from "@/lib/api";
 import LasaConfirmation from "@/components/scanner/LasaConfirmation";
-import GenericAlternativeCard from "@/components/GenericAlternativeCard";
-import { fetchGenericAlternatives } from "@/lib/api/alternatives";
 import { BarcodeScanner } from "@/components/scanner/BarcodeScanner";
 import LazyImage from "@/components/LazyImage";
-import Tesseract from "tesseract.js";
-import {
-    extractExpiryDate,
-    extractBatchNumber,
-    extractMedicineName,
-} from "@/src/utils/medicineParser";
 import { useOfflineStatus } from "@/hooks/useOfflineStatus";
 import { useTranslations } from "next-intl";
 import { buildVerificationShareText, type VerificationShareCopy } from "@/lib/verificationShare";
-import { CopyButton } from "@/components/ui/CopyButton";
-import { structuredLog } from "@/lib/structuredLogger";
 import { saveScanHistory } from "@/lib/db/scanHistory";
 import { recordScanHistory } from "@/lib/scanHistoryUtils";
-
-function formatExpiryForBadge(isoDate: string | null | undefined): string | undefined {
-    if (!isoDate) return undefined;
-    const d = new Date(isoDate);
-    if (isNaN(d.getTime())) return undefined;
-    return `${String(d.getUTCMonth() + 1).padStart(2, "0")}/${d.getUTCFullYear()}`;
-}
-
-function expiryToIso(expiryStr: string): string {
-    const [month, year] = expiryStr.split("/");
-    return `${year}-${month.padStart(2, "0")}-01T00:00:00.000Z`;
-}
-
-function CdscoStatusBadge({ status }: { status: string }) {
-    const config: Record<string, { label: string; className: string }> = {
-        approved: {
-            label: "CDSCO Approved",
-            className:
-                "bg-emerald-50 dark:bg-emerald-950/30 text-emerald-700 dark:text-emerald-450 border-emerald-250 dark:border-emerald-900/30",
-        },
-        recalled: {
-            label: "Recalled",
-            className:
-                "bg-amber-50 dark:bg-amber-950/30 text-amber-700 dark:text-amber-450 border-amber-250 dark:border-amber-900/30",
-        },
-        banned: {
-            label: "Banned",
-            className:
-                "bg-red-50 dark:bg-red-950/30 text-red-700 dark:text-red-400 border-red-250 dark:border-red-900/30",
-        },
-    };
-    const c = config[status] ?? {
-        label: status,
-        className:
-            "bg-(--color-surface-muted) text-(--color-text-secondary) border-(--color-border-muted)",
-    };
-    return (
-        <span
-            className={`inline-block rounded-full border px-2.5 py-1 text-xs font-bold ${c.className}`}
-        >
-            {c.label}
-        </span>
-    );
-}
 
 function formatMedicineDetails(medicine: VerifiedMedicine) {
     return [
@@ -130,350 +57,6 @@ async function copyTextToClipboard(text: string) {
 }
 
 // Result views with dark/light mode surface tokens and variables support
-function VerifiedSafeResult({
-    medicine,
-    scanMeta,
-    onScanAgain,
-    onShare,
-    shareLabel,
-}: {
-    medicine: VerifiedMedicine;
-    scanMeta?: {
-        recentScanCount24h: number;
-        recentScanCount7d: number;
-        suspicious: boolean;
-        suspicionReasons: string[];
-    };
-    onScanAgain: () => void;
-    onShare: () => void;
-    shareLabel: string;
-}) {
-    return (
-        <div className="relative w-full max-w-sm overflow-hidden rounded-[2.5rem] border border-(--color-border-muted) bg-(--color-surface-page) p-8 text-(--color-text-primary) shadow-2xl">
-            <div className="absolute top-0 right-0 left-0 h-2 bg-emerald-500"></div>
-            <div className="flex flex-col items-center space-y-4 text-center">
-                <div className="dark:text-emerald-450 flex h-20 w-20 items-center justify-center rounded-full bg-emerald-100 text-emerald-600 shadow-inner dark:bg-emerald-950/30">
-                    <ShieldCheck size={40} strokeWidth={2.5} />
-                </div>
-                <div>
-                    <h3 className="text-2xl font-black tracking-tight">{medicine.brand_name}</h3>
-                    <p className="font-medium text-(--color-text-secondary)">
-                        Verified by CDSCO Database
-                    </p>
-                </div>
-
-                <CdscoStatusBadge status={medicine.cdsco_approval_status} />
-
-                {scanMeta?.suspicious && (
-                    <div className="border-amber-250 flex w-full items-start gap-3 rounded-2xl border bg-amber-50 p-4 text-left dark:border-amber-900 dark:bg-amber-950/20">
-                        <AlertTriangle
-                            size={18}
-                            className="mt-0.5 shrink-0 text-amber-600 dark:text-amber-400"
-                        />
-                        <p className="text-xs leading-relaxed font-medium text-amber-800 dark:text-amber-400">
-                            {scanMeta.suspicionReasons.join(" ")}
-                        </p>
-                    </div>
-                )}
-
-                <div className="grid w-full grid-cols-2 gap-3 pt-2">
-                    <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) p-3">
-                        <span className="block text-[10px] font-bold tracking-wider text-(--color-text-muted) uppercase">
-                            Batch No.
-                        </span>
-                        <div className="flex items-center justify-between gap-1">
-                            <span className="font-bold text-(--color-text-primary)">
-                                {medicine.batch_number}
-                            </span>
-                            <CopyButton
-                                text={formatMedicineDetails(medicine)}
-                                toastMessage="Medicine Details copied!"
-                            />
-                        </div>
-                    </div>
-                    <ExpiryBadge expiryDate={formatExpiryForBadge(medicine.expiry_date)} />
-                </div>
-
-                <div className="grid w-full grid-cols-2 gap-3">
-                    <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) p-3">
-                        <span className="block text-[10px] font-bold tracking-wider text-(--color-text-muted) uppercase">
-                            Manufacturer
-                        </span>
-                        <span className="text-sm font-bold text-(--color-text-primary)">
-                            {medicine.manufacturer}
-                        </span>
-                    </div>
-                    <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) p-3">
-                        <span className="block text-[10px] font-bold tracking-wider text-(--color-text-muted) uppercase">
-                            Generic Name
-                        </span>
-                        <span className="text-sm font-bold text-(--color-text-primary)">
-                            {medicine.generic_name}
-                        </span>
-                    </div>
-                </div>
-
-                {(medicine.cdsco_approval_status === "recalled" ||
-                    medicine.cdsco_approval_status === "banned") && (
-                    <div className="border-amber-250 flex w-full items-start gap-3 rounded-2xl border bg-amber-50 p-4 text-left dark:border-amber-900 dark:bg-amber-950/20">
-                        <AlertTriangle
-                            size={18}
-                            className="mt-0.5 shrink-0 text-amber-600 dark:text-amber-400"
-                        />
-                        <p className="text-xs leading-relaxed font-medium text-amber-800 dark:text-amber-400">
-                            This medicine has been <strong>{medicine.cdsco_approval_status}</strong>{" "}
-                            by CDSCO. Consult your pharmacist before use.
-                        </p>
-                    </div>
-                )}
-
-                {medicine.cdsco_approval_status === "approved" && (
-                    <div className="border-emerald-250 flex w-full items-start gap-3 rounded-2xl border bg-emerald-50 p-4 text-left dark:border-emerald-900 dark:bg-emerald-950/20">
-                        <Info
-                            size={18}
-                            className="dark:text-emerald-450 mt-0.5 shrink-0 text-emerald-600"
-                        />
-                        <p className="text-xs leading-relaxed font-medium text-emerald-800 dark:text-amber-400">
-                            This medicine matches the official records. Always check the physical
-                            seal before use.
-                        </p>
-                    </div>
-                )}
-
-                <ResultActions
-                    onScanAgain={onScanAgain}
-                    onShare={onShare}
-                    shareLabel={shareLabel}
-                />
-            </div>
-        </div>
-    );
-}
-
-function CounterfeitAlertResult({
-    medicine,
-    onScanAgain,
-    onShare,
-    onCopyMedicineDetails,
-    shareLabel,
-    copied,
-}: {
-    medicine: VerifiedMedicine;
-    onScanAgain: () => void;
-    onShare: () => void;
-    onCopyMedicineDetails: () => void;
-    shareLabel: string;
-    copied: boolean;
-}) {
-    return (
-        <div className="relative w-full max-w-sm overflow-hidden rounded-[2.5rem] border border-(--color-border-muted) bg-(--color-surface-page) p-8 text-(--color-text-primary) shadow-2xl">
-            <div className="absolute top-0 right-0 left-0 h-2 bg-red-500"></div>
-            <div className="flex flex-col items-center space-y-4 text-center">
-                <div className="flex h-20 w-20 items-center justify-center rounded-full bg-red-100 text-red-600 shadow-inner dark:bg-red-950/30 dark:text-red-400">
-                    <AlertTriangle size={40} strokeWidth={2.5} />
-                </div>
-                <div>
-                    <h3 className="text-2xl font-black tracking-tight text-red-700 dark:text-red-400">
-                        Counterfeit Alert
-                    </h3>
-                    <p className="font-medium text-(--color-text-secondary)">
-                        {medicine.brand_name}
-                    </p>
-                </div>
-
-                <div className="grid w-full grid-cols-2 gap-3 pt-2">
-                    <div className="border-red-250/30 rounded-2xl border bg-red-500/10 p-3 dark:border-red-900/30">
-                        <span className="block text-[10px] font-bold tracking-wider text-red-400 uppercase dark:text-red-500/80">
-                            Batch No.
-                        </span>
-                        <div className="flex items-center justify-between gap-1">
-                            <span className="font-bold text-red-700 dark:text-red-400">
-                                {medicine.batch_number}
-                            </span>
-                            <button
-                                onClick={onCopyMedicineDetails}
-                                aria-label="Copy medicine details"
-                                title="Copy medicine details"
-                                className={`shrink-0 rounded-lg p-1.5 transition-all duration-200 ${
-                                    copied
-                                        ? "bg-red-100 text-red-600 dark:bg-red-950/30 dark:text-red-400"
-                                        : "bg-(--color-surface-muted) text-(--color-text-muted) hover:bg-(--color-border-muted) hover:text-(--color-text-primary)"
-                                }`}
-                            >
-                                {copied ? <Check size={14} strokeWidth={3} /> : <Copy size={14} />}
-                            </button>
-                        </div>
-                    </div>
-                    <div className="border-red-250/30 rounded-2xl border bg-red-500/10 p-3 dark:border-red-900/30">
-                        <span className="block text-[10px] font-bold tracking-wider text-red-400 uppercase dark:text-red-500/80">
-                            Manufacturer
-                        </span>
-                        <span className="text-sm font-bold text-red-700 dark:text-red-400">
-                            {medicine.manufacturer}
-                        </span>
-                    </div>
-                </div>
-
-                <div className="border-red-250 flex w-full items-start gap-3 rounded-2xl border bg-red-50 p-4 text-left dark:border-red-900 dark:bg-red-950/20">
-                    <AlertTriangle
-                        size={18}
-                        className="mt-0.5 shrink-0 text-red-600 dark:text-red-400"
-                    />
-                    <p className="text-xs leading-relaxed font-bold text-red-800 dark:text-red-400">
-                        WARNING: This medicine has been flagged as counterfeit. Do NOT consume.
-                        Report to your nearest pharmacy or call the CDSCO helpline immediately.
-                    </p>
-                </div>
-
-                <ResultActions
-                    onScanAgain={onScanAgain}
-                    onShare={onShare}
-                    shareLabel={shareLabel}
-                />
-            </div>
-        </div>
-    );
-}
-
-function UnverifiedResult({
-    brandName,
-    batchNumber,
-    expiryDate,
-    onScanAgain,
-    onShare,
-    shareLabel,
-}: {
-    brandName?: string;
-    batchNumber?: string;
-    expiryDate?: string;
-    onScanAgain: () => void;
-    onShare: () => void;
-    shareLabel: string;
-}) {
-    return (
-        <div className="relative w-full max-w-sm overflow-hidden rounded-[2.5rem] border border-(--color-border-muted) bg-(--color-surface-page) p-8 text-(--color-text-primary) shadow-2xl">
-            <div className="absolute top-0 right-0 left-0 h-2 bg-amber-500"></div>
-            <div className="flex flex-col items-center space-y-4 text-center">
-                <div className="dark:text-amber-450 flex h-20 w-20 items-center justify-center rounded-full bg-amber-100 text-amber-600 shadow-inner dark:bg-amber-950/30">
-                    <XCircle size={40} strokeWidth={2.5} />
-                </div>
-                <div>
-                    <h3 className="dark:text-amber-450 text-2xl font-black tracking-tight text-amber-700">
-                        {brandName || "Unverified Medicine"}
-                    </h3>
-                    <p className="font-medium text-(--color-text-secondary)">
-                        No match found in CDSCO Database
-                    </p>
-                </div>
-
-                {(batchNumber || expiryDate) && (
-                    <div className="grid w-full grid-cols-2 gap-3 pt-2">
-                        <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) p-3">
-                            <span className="block text-[10px] font-bold tracking-wider text-(--color-text-muted) uppercase">
-                                Batch No.
-                            </span>
-                            <span className="font-bold text-(--color-text-primary)">
-                                {batchNumber || "Unknown"}
-                            </span>
-                        </div>
-                        <ExpiryBadge expiryDate={expiryDate} />
-                    </div>
-                )}
-
-                <div className="border-amber-250 flex w-full items-start gap-3 rounded-2xl border bg-amber-50 p-4 text-left dark:border-amber-900 dark:bg-amber-950/20">
-                    <Info
-                        size={18}
-                        className="mt-0.5 shrink-0 text-amber-600 dark:text-amber-400"
-                    />
-                    <p className="dark:text-amber-450 text-xs leading-relaxed font-medium text-amber-800">
-                        No matching record was found for this medicine batch in the CDSCO database.
-                        Please verify the spelling or report it if suspicious.
-                    </p>
-                </div>
-
-                <ResultActions
-                    onScanAgain={onScanAgain}
-                    onShare={onShare}
-                    shareLabel={shareLabel}
-                />
-            </div>
-        </div>
-    );
-}
-
-function ErrorResult({
-    message,
-    onRetry,
-    isOffline,
-}: {
-    message: string;
-    onRetry: () => void;
-    isOffline?: boolean;
-}) {
-    return (
-        <div className="relative w-full max-w-sm overflow-hidden rounded-[2.5rem] border border-(--color-border-muted) bg-(--color-surface-page) p-8 text-(--color-text-primary) shadow-2xl">
-            <div className="absolute top-0 right-0 left-0 h-2 bg-slate-400"></div>
-            <div className="flex flex-col items-center space-y-4 text-center">
-                <div className="flex h-20 w-20 items-center justify-center rounded-full bg-(--color-surface-muted) text-(--color-text-secondary) shadow-inner">
-                    <AlertCircle size={40} strokeWidth={2.5} />
-                </div>
-                <div>
-                    <h3 className="text-2xl font-black tracking-tight text-(--color-text-primary)">
-                        {isOffline ? "Connection Lost" : "Verification Failed"}
-                    </h3>
-                    <p className="text-sm font-medium whitespace-pre-wrap text-slate-500">
-                        {message}
-                    </p>
-                </div>
-
-                <button
-                    onClick={onRetry}
-                    disabled={isOffline}
-                    className="w-full rounded-2xl bg-slate-900 py-4 font-bold text-white shadow-lg shadow-slate-900/20 transition-all hover:bg-slate-800 disabled:cursor-not-allowed disabled:opacity-50 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200"
-                >
-                    {isOffline ? "Waiting for connection..." : "Try Again"}
-                </button>
-            </div>
-        </div>
-    );
-}
-
-function ResultActions({
-    onScanAgain,
-    onShare,
-    shareLabel,
-}: {
-    onScanAgain: () => void;
-    onShare: () => void;
-    shareLabel: string;
-}) {
-    return (
-        <div className="no-print grid w-full grid-cols-1 gap-3">
-            <button
-                onClick={onScanAgain}
-                className="w-full rounded-2xl bg-slate-900 py-4 font-bold text-white shadow-lg shadow-slate-900/20 transition-all hover:bg-slate-800 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200"
-            >
-                Scan Another
-            </button>
-            <div className="grid grid-cols-2 gap-3">
-                <Link
-                    href="/"
-                    className="flex items-center justify-center gap-2 rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) py-3.5 font-semibold text-(--color-text-primary) transition-all hover:bg-(--color-border-muted)"
-                >
-                    <Home size={18} />
-                    <span>Home</span>
-                </Link>
-                <button
-                    onClick={onShare}
-                    className="flex items-center justify-center gap-2 rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) py-3.5 font-semibold text-(--color-text-primary) transition-all hover:bg-(--color-border-muted)"
-                >
-                    <Share2 size={18} />
-                    <span>{shareLabel}</span>
-                </button>
-            </div>
-        </div>
-    );
-}
 
 export default function ScanPage() {
     const tScan = useTranslations("Scan");
@@ -483,26 +66,45 @@ export default function ScanPage() {
     const { isOffline, registerRetryCallback, unregisterRetryCallback } = useOfflineStatus();
     const abortControllerRef = useRef<AbortController | null>(null);
     const isMountedRef = useRef(true);
-
     const [isScanning, setIsScanning] = useState(false);
-    const [showResult, setShowResult] = useState(false);
-    const [uploadedImage, setUploadedImage] = useState<string | null>(null);
     const [copied, setCopied] = useState(false);
     const [batchInput, setBatchInput] = useState("");
-    const [verifyResult, setVerifyResult] = useState<VerifyResult | null>(null);
-    const [verifyError, setVerifyError] = useState<string | null>(null);
-    const [ocrText, setOcrText] = useState<string | null>(null);
-    const [ocrConfidence, setOcrConfidence] = useState<number | null>(null);
-    const [parsedBrand, setParsedBrand] = useState<string>("");
-    const [parsedBatch, setParsedBatch] = useState<string>("");
-    const [parsedExpiry, setParsedExpiry] = useState<string>("");
     const [isCameraActive, setIsCameraActive] = useState(false);
-    const [ocrStatus, setOcrStatus] = useState<
-        "idle" | "scanning-barcode" | "extracting-text" | "done" | "error"
-    >("idle");
 
-    const ocrWorkerRef = useRef<Tesseract.Worker | null>(null);
-    const ocrCancelledRef = useRef(false);
+    const [showResult, setShowResult] = useState(false);
+    const {
+        verifyResult,
+        verifyError,
+        lasaMatches,
+        showLasaConfirmation,
+        pendingVerifyResult,
+
+        setVerifyResult,
+        setVerifyError,
+        setPendingVerifyResult,
+        setShowLasaConfirmation,
+
+        handleVerify,
+        processVerificationResult,
+    } = useMedicineVerification(abortControllerRef, isMountedRef, setIsScanning, setShowResult);
+    const {
+        uploadedImage,
+        ocrText,
+        ocrConfidence,
+        parsedBrand,
+        parsedBatch,
+        parsedExpiry,
+        handleFileUpload,
+        reset,
+    } = useMedicineImageUpload({
+        handleVerify,
+        processVerificationResult,
+        setVerifyError,
+        setShowResult,
+        setBatchInput,
+        setVerifyResult,
+        setIsScanning,
+    });
 
     // Auto-retry when coming back online
     const handleVerifyRef = useRef<(batch: string) => Promise<void>>(null as any);
@@ -521,7 +123,6 @@ export default function ScanPage() {
 
         return () => {
             isMountedRef.current = false;
-            ocrCancelledRef.current = true;
             if (abortControllerRef.current) {
                 abortControllerRef.current.abort();
             }
@@ -529,20 +130,7 @@ export default function ScanPage() {
         };
     }, [showResult, verifyError, batchInput, registerRetryCallback, unregisterRetryCallback]);
 
-    useEffect(() => {
-        return () => {
-            console.log("Tesseract worker terminated on ScanPage unmount");
-            if (ocrWorkerRef.current) {
-                ocrWorkerRef.current.terminate();
-                ocrWorkerRef.current = null;
-            }
-        };
-    }, []);
-
     // LASA Check State
-    const [lasaMatches, setLasaMatches] = useState<LasaMatch[]>([]);
-    const [showLasaConfirmation, setShowLasaConfirmation] = useState(false);
-    const [pendingVerifyResult, setPendingVerifyResult] = useState<VerifyResult | null>(null);
     const shareCopy: VerificationShareCopy = {
         realStatus: tScan("share.real_status"),
         suspiciousStatus: tScan("share.suspicious_status"),
@@ -556,84 +144,6 @@ export default function ScanPage() {
         unknownBatch: tScan("share.unknown_batch"),
         unknownManufacturer: tScan("share.unknown_manufacturer"),
     };
-
-    // Alternative Finder State
-    const [alternativeData, setAlternativeData] = useState<any | null>(null);
-    const [loadingAlternative, setLoadingAlternative] = useState<boolean>(false);
-
-    const fetchAlternativesForMedicine = useCallback(async (medicineNameOrId: string) => {
-        setLoadingAlternative(true);
-        setAlternativeData(null);
-        try {
-            const data = await fetchGenericAlternatives(medicineNameOrId);
-            setAlternativeData(data);
-        } catch (err) {
-            console.error("Failed to fetch alternatives:", err);
-        } finally {
-            setLoadingAlternative(false);
-        }
-    }, []);
-
-    const processVerificationResult = useCallback(
-        async (result: VerifyResult, fallbackBrandName?: string) => {
-            if (!result.verified) {
-                setVerifyResult(result);
-                setAlternativeData(null);
-
-                void recordScanHistory(result, fallbackBrandName).catch((error) => {
-                    console.error("Failed to save scan history:", error);
-                });
-
-                setShowResult(true);
-
-                return;
-            }
-            try {
-                const medicineName = result.medicine.brand_name || fallbackBrandName;
-                if (!medicineName) {
-                    setVerifyResult(result);
-                    setAlternativeData(null);
-
-                    void recordScanHistory(result, fallbackBrandName).catch((error) => {
-                        console.error("Failed to save scan history:", error);
-                    });
-
-                    setShowResult(true);
-
-                    return;
-                }
-
-                // Start fetching generic alternatives
-                void fetchAlternativesForMedicine(medicineName);
-
-                const lasaRes = await checkLasaConflicts(medicineName);
-                if (lasaRes.hasConflicts && lasaRes.matches.length > 0) {
-                    setLasaMatches(lasaRes.matches);
-                    setPendingVerifyResult(result);
-                    setShowLasaConfirmation(true);
-                    setShowResult(true);
-                } else {
-                    setVerifyResult(result);
-
-                    void recordScanHistory(result, fallbackBrandName).catch((error) => {
-                        console.error("Failed to save scan history:", error);
-                    });
-
-                    setShowResult(true);
-                }
-            } catch (error) {
-                console.error("LASA check error:", error);
-                setVerifyResult(result);
-
-                void recordScanHistory(result, fallbackBrandName).catch((historyError) => {
-                    console.error("Failed to save scan history:", historyError);
-                });
-
-                setShowResult(true);
-            }
-        },
-        [fetchAlternativesForMedicine]
-    );
 
     const handleConfirmScanned = () => {
         if (pendingVerifyResult) {
@@ -662,7 +172,6 @@ export default function ScanPage() {
         try {
             const brandRes = await verifyMedicineByBrand(conflictName, controller.signal);
             if (!isMountedRef.current || controller.signal.aborted) return;
-            setParsedBrand(conflictName);
             await processVerificationResult(brandRes, conflictName);
         } catch (err) {
             if (!isMountedRef.current || controller.signal.aborted) return;
@@ -686,54 +195,6 @@ export default function ScanPage() {
             }
         }
     };
-
-    const handleVerify = useCallback(
-        async (batch: string) => {
-            if (!batch.trim()) {
-                toast.error("Please enter a batch number to verify");
-                return;
-            }
-            const normalizedBatch = batch.trim();
-
-            if (abortControllerRef.current) {
-                abortControllerRef.current.abort();
-            }
-            const controller = new AbortController();
-            abortControllerRef.current = controller;
-
-            setIsScanning(true);
-            setShowResult(false);
-            setVerifyResult(null);
-            setVerifyError(null);
-
-            try {
-                const result = await verifyMedicine(normalizedBatch, controller.signal);
-                if (!isMountedRef.current || controller.signal.aborted) return;
-                await processVerificationResult(result, undefined);
-            } catch (err) {
-                if (!isMountedRef.current || controller.signal.aborted) return;
-                const errorMsg = err instanceof Error ? err.message : "Verification failed";
-                if (errorMsg === "Request was cancelled.") {
-                    return;
-                }
-                setVerifyError(errorMsg);
-                void saveScanHistory({
-                    id: crypto.randomUUID(),
-                    timestamp: Date.now(),
-                    medicineName: batch.trim() || "Unknown Medicine",
-                    status: "SUSPICIOUS",
-                }).catch((error) => {
-                    console.error("Failed to save scan history:", error);
-                });
-                setShowResult(true);
-            } finally {
-                if (isMountedRef.current && !controller.signal.aborted) {
-                    setIsScanning(false);
-                }
-            }
-        },
-        [processVerificationResult]
-    );
 
     // Keep handleVerifyRef current
     useEffect(() => {
@@ -759,288 +220,6 @@ export default function ScanPage() {
         }
     }, [verifyResult]);
 
-    const MAX_FILE_SIZE = 10 * 1024 * 1024;
-    const COMPRESSION_THRESHOLD = 2 * 1024 * 1024;
-
-    const handleFileUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
-        const file = e.target.files?.[0];
-        if (!file) return;
-
-        let processedFile = file;
-
-        if (file.size > MAX_FILE_SIZE) {
-            toast.error("File exceeds 10MB limit");
-            e.target.value = "";
-            return;
-        }
-
-        if (file.size > COMPRESSION_THRESHOLD) {
-            try {
-                processedFile = await imageCompression(file, {
-                    maxSizeMB: 1,
-                    maxWidthOrHeight: 1920,
-                    useWebWorker: true,
-                });
-            } catch {
-                toast.error("Failed to compress image");
-                return;
-            }
-        }
-
-        const reader = new FileReader();
-        let dataUrl: string;
-        try {
-            dataUrl = await new Promise<string>((resolve, reject) => {
-                reader.onloadend = () => resolve(reader.result as string);
-                reader.onerror = () => reject(new Error("Failed to read image file"));
-                reader.readAsDataURL(processedFile);
-            });
-        } catch {
-            toast.error("Could not read the image file. Please try again.");
-            e.target.value = "";
-            return;
-        }
-        setUploadedImage(dataUrl);
-        e.target.value = "";
-
-        if (abortControllerRef.current) {
-            abortControllerRef.current.abort();
-        }
-        const controller = new AbortController();
-        abortControllerRef.current = controller;
-
-        setIsScanning(true);
-        setShowResult(false);
-        setVerifyResult(null);
-        setVerifyError(null);
-        setOcrText(null);
-        setOcrConfidence(null);
-        setParsedBrand("");
-        setParsedBatch("");
-        setParsedExpiry("");
-
-        ocrCancelledRef.current = false;
-
-        try {
-            // ── Step 1: Try ZXing barcode decode from uploaded image ──────────
-            setOcrStatus("scanning-barcode");
-            let barcodeFound = false;
-
-            try {
-                const { BrowserMultiFormatReader } = await import("@zxing/browser");
-                const { DecodeHintType, BarcodeFormat } = await import("@zxing/library");
-
-                const hints = new Map();
-                hints.set(DecodeHintType.POSSIBLE_FORMATS, [
-                    BarcodeFormat.CODE_128,
-                    BarcodeFormat.QR_CODE,
-                    BarcodeFormat.EAN_13,
-                    BarcodeFormat.EAN_8,
-                    BarcodeFormat.CODE_39,
-                    BarcodeFormat.DATA_MATRIX,
-                ]);
-                hints.set(DecodeHintType.TRY_HARDER, true);
-
-                const reader = new BrowserMultiFormatReader(hints);
-                const zxingResult = await reader.decodeFromImageUrl(dataUrl);
-                const barcodeText = zxingResult.getText().trim();
-                if (barcodeText) {
-                    barcodeFound = true;
-                    if (!isMountedRef.current || controller.signal.aborted) return;
-                    setBatchInput(barcodeText);
-                    setOcrStatus("done");
-                    toast.success(`Barcode detected: ${barcodeText} — verifying…`);
-                    await handleVerify(barcodeText);
-                    return;
-                }
-            } catch (error) {
-                structuredLog({
-                    log_level: "warn",
-                    route: "/scan",
-                    meta: {
-                        message: "[scan] Barcode detection (ZXing) failed, falling back to OCR",
-                        error: error instanceof Error ? error.message : String(error),
-                    },
-                });
-            }
-
-            if (!isMountedRef.current || controller.signal.aborted) return;
-            if (barcodeFound || ocrCancelledRef.current) return;
-
-            // ── Step 2: Tesseract.js OCR Fallback ────────────────────────────
-            setOcrStatus("extracting-text");
-
-            if (!ocrWorkerRef.current) {
-                ocrWorkerRef.current = await Tesseract.createWorker("eng");
-            }
-
-            if (!isMountedRef.current || controller.signal.aborted || ocrCancelledRef.current)
-                return;
-
-            const timeoutPromise = new Promise<never>((_, reject) => {
-                setTimeout(() => reject(new Error("OCR timed out")), 30000);
-            });
-
-            const ocrPromise = ocrWorkerRef.current.recognize(dataUrl);
-            const { data } = await Promise.race([ocrPromise, timeoutPromise]);
-
-            if (!isMountedRef.current || controller.signal.aborted || ocrCancelledRef.current)
-                return;
-
-            const rawText = data.text;
-            if (!rawText || !rawText.trim()) {
-                toast.warning("No clear text found in image.");
-                setVerifyError(
-                    "Failed to read medicine text. Please ensure the image is clear or upload another one."
-                );
-                setOcrStatus("error");
-                setShowResult(true);
-                setIsScanning(false);
-                return;
-            }
-
-            setOcrText(rawText);
-            setOcrConfidence(data.confidence / 100);
-            setOcrStatus("done");
-            toast.success("OCR extraction complete!");
-
-            // Parse OCR Text using utility regex
-            const parsedBatchNum = extractBatchNumber(rawText);
-            const parsedExpiryStr = extractExpiryDate(rawText);
-            const medName = extractMedicineName(rawText);
-
-            if (parsedBatchNum) setParsedBatch(parsedBatchNum);
-            if (parsedExpiryStr) setParsedExpiry(parsedExpiryStr);
-            if (medName) setParsedBrand(medName);
-
-            if (parsedBatchNum) {
-                setBatchInput(parsedBatchNum);
-            }
-
-            // Database Lookup Strategy
-            let finalResult: VerifyResult | null = null;
-
-            if (parsedBatchNum) {
-                try {
-                    const batchRes = await verifyMedicine(parsedBatchNum, controller.signal);
-                    if (batchRes.verified) {
-                        finalResult = batchRes;
-                    }
-                } catch (error) {
-                    structuredLog({
-                        log_level: "warn",
-                        route: "/scan",
-                        meta: {
-                            message: "[scan] Batch verification failed, trying brand match",
-                            batch: parsedBatchNum,
-                            error: error instanceof Error ? error.message : String(error),
-                        },
-                    });
-                }
-            }
-
-            if (!isMountedRef.current || controller.signal.aborted) return;
-
-            if (!finalResult && medName) {
-                try {
-                    const matchRes = await fuzzyMatchBrand(medName, controller.signal);
-                    if (matchRes && matchRes.length > 0) {
-                        const topMatch = matchRes[0];
-                        if (topMatch.score >= 60) {
-                            setParsedBrand(topMatch.name);
-                            const brandRes = await verifyMedicineByBrand(
-                                topMatch.name,
-                                controller.signal
-                            );
-                            if (brandRes.verified) {
-                                finalResult = brandRes;
-                            }
-                        }
-                    }
-                } catch (error) {
-                    structuredLog({
-                        log_level: "warn",
-                        route: "/scan",
-                        meta: {
-                            message: "[scan] Fuzzy brand match verification failed",
-                            brand: medName,
-                            error: error instanceof Error ? error.message : String(error),
-                        },
-                    });
-                }
-            }
-
-            if (!isMountedRef.current || controller.signal.aborted) return;
-
-            if (finalResult && finalResult.verified) {
-                const updatedMedicine = { ...finalResult.medicine };
-                if (parsedBatchNum) {
-                    updatedMedicine.batch_number = parsedBatchNum;
-                }
-                if (parsedExpiryStr) {
-                    updatedMedicine.expiry_date = expiryToIso(parsedExpiryStr);
-                }
-                await processVerificationResult(
-                    { verified: true, medicine: updatedMedicine },
-                    parsedBrand
-                );
-            } else {
-                const unverifiedResult =
-                    finalResult ||
-                    ({
-                        verified: false,
-                        message: "No match found in CDSCO Database",
-                    } satisfies VerifyResult);
-                await processVerificationResult(
-                    unverifiedResult,
-                    parsedBrand || medName || undefined
-                );
-            }
-        } catch (err) {
-            if (!isMountedRef.current || controller.signal.aborted || ocrCancelledRef.current)
-                return;
-
-            if (ocrWorkerRef.current) {
-                await ocrWorkerRef.current.terminate();
-                ocrWorkerRef.current = null;
-            }
-
-            const errorMsg = err instanceof Error ? err.message : String(err);
-            if (errorMsg === "OCR timed out") {
-                toast.error("OCR timed out. Please try again with a clearer image.");
-                setVerifyError(
-                    "The scan took too long. Please ensure the image is clear and try again."
-                );
-                void saveScanHistory({
-                    id: crypto.randomUUID(),
-                    timestamp: Date.now(),
-                    medicineName: parsedBrand || parsedBatch || "Unknown Medicine",
-                    status: "SUSPICIOUS",
-                }).catch((error) => {
-                    console.error("Failed to save scan history:", error);
-                });
-            } else {
-                toast.error("Failed to extract text from image.");
-                setVerifyError(
-                    "Unable to read text from this image. Please try a clearer photo or enter the batch number manually."
-                );
-                void saveScanHistory({
-                    id: crypto.randomUUID(),
-                    timestamp: Date.now(),
-                    medicineName: parsedBrand || parsedBatch || "Unknown Medicine",
-                    status: "SUSPICIOUS",
-                }).catch((error) => {
-                    console.error("Failed to save scan history:", error);
-                });
-            }
-            setOcrStatus("error");
-        } finally {
-            if (isMountedRef.current && !controller.signal.aborted && !ocrCancelledRef.current) {
-                setIsScanning(false);
-                setShowResult(true);
-            }
-        }
-    };
     const handleBarcodeScan = useCallback(
         async (scannedText: string) => {
             setIsVerifying(true);
@@ -1057,39 +236,20 @@ export default function ScanPage() {
         [handleVerify]
     );
 
-    const handleScanAgain = async () => {
-        if (ocrWorkerRef.current) {
-            await ocrWorkerRef.current.terminate();
-            ocrWorkerRef.current = null;
-        }
-        ocrCancelledRef.current = true;
-        setIsScanning(false);
+    const handleScanAgain = () => {
+        reset();
         setShowResult(false);
-        setUploadedImage(null);
         setVerifyResult(null);
         setVerifyError(null);
         setBatchInput("");
-        setOcrText(null);
-        setOcrConfidence(null);
-        setParsedBrand("");
-        setParsedBatch("");
-        setParsedExpiry("");
         setIsCameraActive(false);
-        setOcrStatus("idle");
     };
 
-    const handleDismissResult = async () => {
-        if (ocrStatus === "error" && ocrWorkerRef.current) {
-            await ocrWorkerRef.current.terminate();
-            ocrWorkerRef.current = null;
-        }
+    const handleDismissResult = () => {
+        reset();
         setShowResult(false);
         setVerifyResult(null);
         setVerifyError(null);
-        setParsedBrand("");
-        setParsedBatch("");
-        setParsedExpiry("");
-        setOcrStatus("idle");
     };
 
     const handleShare = async () => {
@@ -1235,28 +395,15 @@ export default function ScanPage() {
                                 {!verifyError &&
                                     verifyResult?.verified &&
                                     !verifyResult.medicine.is_counterfeit_alert && (
-                                        <div className="flex w-full max-w-sm flex-col items-center gap-6">
-                                            <VerifiedSafeResult
-                                                medicine={verifyResult.medicine}
-                                                scanMeta={verifyResult.scanMeta}
-                                                onScanAgain={handleScanAgain}
-                                                onShare={handleShare}
-                                                shareLabel={tScan("share.button")}
-                                            />
-                                            {loadingAlternative && (
-                                                <div className="flex w-full items-center justify-center rounded-[2.5rem] border border-(--color-border-muted) bg-slate-50 p-6 dark:bg-slate-900">
-                                                    <div className="h-6 w-6 animate-spin rounded-full border-2 border-emerald-500 border-t-transparent"></div>
-                                                    <span className="ml-3 text-xs font-bold text-emerald-600 dark:text-emerald-400">
-                                                        Finding cheaper alternatives...
-                                                    </span>
-                                                </div>
-                                            )}
-                                            {!loadingAlternative && alternativeData && (
-                                                <GenericAlternativeCard
-                                                    alternative={alternativeData}
-                                                />
-                                            )}
-                                        </div>
+                                        <VerifiedSafeResult
+                                            medicine={verifyResult.medicine}
+                                            scanMeta={verifyResult.scanMeta}
+                                            onScanAgain={handleScanAgain}
+                                            onShare={handleShare}
+                                            onCopyMedicineDetails={handleCopyMedicineDetails}
+                                            shareLabel={tScan("share.button")}
+                                            copied={copied}
+                                        />
                                     )}
                                 {!verifyError && verifyResult && !verifyResult.verified && (
                                     <UnverifiedResult
